requirements.txt
pytchat
google.generativeai
pydub
pyttsx3
simpleaudio
pyaudio
ffmpeg
